Decouple file & document logic
